Move AboutHairBox into AboutMeBox.
